auth: plug isChallengeDomain security hole
Pat Dowler pointed out that testing only challenge parameters like
the access_url to determine whether a challenge matches an existing
AuthContext is not secure, since an evil endpoint could supply another
service's access_url in order to obtain authentication information such
as a bearer token owned by a context established for the other service.
So modify the isChallengeDomain javadocs to point this out, and restrict
the existing implementations accordingly.

Even with this adjustment, BearerIvoaAuthScheme cannot be fixed up
so that it is (a) safe and (b) capable of using an endpoint in a
different domain than the one where the resources of interest live.
Since the only service currently offering ivoa_bearer is CADC,
where the login node is in a different place than the TAP service,
that makes it a bit useless.  So BearerIvoaAuthScheme is no longer
used by AuthManager.  The code is kept here in case it becomes
useful at some point in the future.