ttools: random(), randomGaussian() replaces $random
The implementation $random special token is fatally flawed;
it will not always generate the same value for the same table cell,
because of difficulties in trying to work out what seed it should use.

Therefore deprecate it and provide instead a random(seed) function
in class Maths.
At the same time, add a corresponding randomGaussian() function.