vo: implement UNION-based RegTAP keyword query
Implement two different options for the RegTAP createKeywordQuery method,
the original one based on ORs and a new one based on
ivoid IN (SELECT ... UNION ... SELECT ...).
The latter is much faster for Postgres (DaCHS) services, but won't work
unless the optional ADQL 2.1 language feature UNION is present.

At this commit the new option is unused, so behaviour is unchanged.