vo: use UNION-based keyword RegTAP query for DaCHS
The UNION-based option is now used for some RegTAP services that
are known to support it, namely services recognised as DaCHS TAP
service URLs.

Hard-coding the known UNION-supporting RegTAP services like this is
clearly hacky, but it will probably do the right thing in nearly all
cases at time of writing, since the TAP service URLs are provided
from a pre-populated selector, and right now the only RegTAP services
supporting UNION are DaCHS.
The Right Thing in principle would be to query the TAPRegExt
capabilities to see if UNION support is present but
(a) this presents practical difficulties in the implementation since it
requires an additional service call (solvable but requires reorganisation)
and (b) I wouldn't necessarily trust the ADQL I'm using to work
well/at all on non-DaCHS RegTAP services even if they do support UNION;
I'd want to test them first.