pds: add test to ensure all FieldTypes are handled
This test would have complained about the missing ASCII_BOOLEAN
FieldReader if it had been in place.  Since it passes now, it
demonstrates that all the FieldType enum values are catered for
by the FieldReader class.

Note that failure of this test is achieved by a runtime Java assert
statement (not a JUnit assert method invocation), which means that the
build file needs to make sure tests are done not only with
-enableassertions, but also in a forked JVM, so that assertion failures
translate to JUnit failures.