topcat: remove Combine key from Densogram docs
The Combine key is not present in Densogram configuration,
but it was documented in SUN/253.
I think it was present at some point, but got removed because
many of the combine options didn't make sense (as for the KNN).
That *might* not be true, but this commit at any rate brings the
documentation in line with the behaviour.