topcat: provide URL in activation messages
Take steps to ensure that the Outcome message resulting from
Activator.activateRow contains the text of the URL that it
refers to, for URL-based activation types.
Outcome messages are not supposed to contain context, to keep them
short and since it can be provided elsewhere, but in the
ActivationWindow the URL typically is not provided elsewhere,
so add machinery to make sure the URL is inserted into the
Outcome message if it's not already present.