ttools: add TFCAT AreaMapper in AreaDomain
AreaDomain can now decode areas specified as TFCat texts.
Support is currently incomplete, in particular CRS is ignored
and holes in (Multi)Polygons are not treated correctly.
It's also not been tested much.  These may get fixed in the future
if it turns out people actually want to use TFCat descriptions
in TOPCAT/STILTS, but for now it's mostly there as a marker and
proof of concept.