topcat: fix JEL exception handling regression bug
If a RuntimeException was thrown when an expression was evaluated
by a TopcatJELEvaluator it was getting rethrown as an IOException
which is not always caught.  This meant that plots stopped working
(with a WARNING level stack trace) when such an expression was
evaluated.  This is bad; such erroneous evaluations are quite
common and should just be interpreted as null results
(even though they are much better avoided for performance reasons).
Fix it so that the evaluations throw RuntimeExceptions as themselves
rather than repackaging them as IOExceptions.

This regression bug was introduced at fc57048e (a year ago) when
TopcatJELEvaluator was brought in to improve evaluation efficiency.