ttools: add resource checking MOC coverage calculations
Add some code in MocCoverage.discOverlaps to defend against possible
resource exhaustion.  If a test is made to identify overlap of a
disc which is very large compared to the maximum order of a given MOC,
the HealpixNestedBMOC.overlappingCells call can require a very
large amount of memory (to accommodate all the little tiles around
the edge of the disc) and potentially fail with an OutOfMemoryError.
So limit the resolution of the disc to prevent that happening.

This was causing trouble in the presence of some MOCs from the CDS
MOCServer that had anomalously high maximum depths (28).  Those have
now been fixed, but this change should defend against similar issues
in the future.