votable: use DataBufferedI/OStream where possible
Avoid use of java.io.DataInputStream/DataOutputStream
(even when overlaid on a BufferedI/OStream) in favour of
uk.ac.starlink.util.DataBufferedI/OStream, which can be
significantly faster.

For example, when writing to fits-plus or
votable(format=BINARY2,inline=false) write speed is faster by better
than a factor of 2.  For output to binary-inline the effect is
smaller because of inefficiencies in the Base64 output.

BinaryRowSequence also no longer uses a PushbackInputStream;
as far as I can tell it can do without it.