votable: improve performance of Base64 de/coding
Look at the Base64 encoding and decoding in VOTable (used mainly for
inline BINARY/BINARY2 data streams) to see if there are performance
issues that can be improved.

Up till now I have used uk.ac.starlink.util.Base64{Input/Output}Stream,
which were introduced when there was no Base64 encoding in the J2SE.
At java 1.8, which is now the target platform, the java.util.Base64
class was introduced which is capable of doing this.
Benchmarking however does not support the idea of ditching the
starlink.util classes in favour of the java.util ones.
For input, the java.util class is faster, but for output the
starlink.util one is, though this is also affected by the underlying
data streams; moreover the API of java.util.Base64.Encoder makes
it hard to write base64 into a stream without closing it at the
end of the base64 section (which we need to do for VOTable).

To summarise the effects of these changes: reading inline BINARY
tables is now faster by about a factor of 4.  Not much change in
writing performance.