ttools: replace StringIO with BytesIO in jython test
The jython StringIO class seems to behave strangely on occasion
when receiving byte arrays, probably trying to be clever with Unicode.
Replace it with BytesIO which is a more appropriate class
(I don't think it existed in the Jython version I was originally using).
No change in behaviour here, but it prevents a unit test failure
appearing following later changes to the FITS I/O implementation.