fits: cope with spaces in TDIMn headers
If the TDIMn header for a FITS BINTABLE column contained embedded
spaces (e.g. "(30, 30)" instead of "(30,30)") the parsing failed
and gave a partially filled dimensions array.  This generally ended
up discarding the shape information for the column.

The FITS standard (FITS 4.0 sec 7.3.2) doesn't say that spaces are
permitted here, so this usage is probably not legal FITS, but it
exists in at least some cases, so it makes sense to cater for it.
Now fixed.