table: "-" in 1-character MRT field no longer null
Adjust the null handling in MRT files so that a "-" in a
single-character field is no longer interpreted as a null,
in consideration of "DE-"-type fields giving latitude sign values.
There's no well-defined MRT standard so the Right behaviour is
not clear here, but this agreed in discussion with Greg Schwarz (AAS)
as reasonable.