ttools: taplint survives missing capabilities
Fix taplint so that if the capabilities is absent at startup it
reports that fact through the reporter rather than failing.
The report is in a phantom ad hoc stage PRE, which is a bit hacky,
but it's better than the previous behaviour, especially when
harnessing the application for automated validation of multiple services.