ttools: add new task xsdvalidate
It validates against XSD schemas as declared in a target XML document,
allowing custom specification of schema locations as required.

Not very stiltsy, but a useful item to have, and added at request of
Markus Demleitner who says there are not many other suitable tools
out there to do this.