jsamp: remove org.xml.sax classes from xmlrpc jar file
The xmlrpc third party library jar files in the JSAMP and PLASTIC
packages both contained classes in the org.xml.sax namespace.
Since classes with these names are present in the J2SE,
it's not a good idea to have them there.  Must be a remnant of a
time before they were in the J2SE.

I didn't check whether there were API or implementation differences
between the bundled versions and the J2SE versions, but since
nothing's broken before now, I'm guessing not.  This change is
unlikely to have a material effect on anything, but slims down
the dependencies very slightly.