fits: interpret (a,b,c) BINTABLE header values as arrays
If a FITS header is going to be used as a table parameter for a BINTABLE,
and it has the form (x, x, x, ...) and can be parsed as a numeric array,
parse it thus and make it available in its array form rather than the
raw string form in the StarTable output.

This is a bit hacky, but it does what's wanted for Gaia XP spectra
which are serialized to FITS BINTABLEs with a header that looks like

   SAMPLING= '(336.0, 338.0, 340.0, ...)'

(usually making use of CONTINUE headers).