tjoin: provide threading control for RowMatcher
The RowMatcher factory method now takes a RowRunner parameter to
control how matching is done (null corresponds to the legacy
sequential code).  Corresponding UI changes are made in the
STILTS and TOPCAT application code to make use of this.

At present, the legacy sequential code is the default,
given that testing of the multithreaded implementation
has not yet been exhaustive, and errors are a realistic possibility.
The STILTS documentation reflects this.