util: retire legacy Base64OutputStream
The original Base64OutputStream implementation is no longer used,
it's replaced by the much faster BufferedBase64OutputStream.
While the latter is more or less a drop-in replacement,
I retain the name change to make the behaviour clear.

The old class is retained in the unit tests only, and renamed
LegacyBase64OutputStream, as an extra test that the buffered
implementation is behaving properly (though that's not really
necessary given the tests already compare with java.util.Base64).