ttools: document "Aitoff" projection(s) as "Hammer-Aitoff"
The sky projection I have been using forever in the sky plots under the
name "Aitoff" (and "Aitoff0") are, it turns out, actually Hammer-Aitoff.
This commit doesn't change the names, but it clarifies in the documentation
that this is the case.

The relevant code is from a skyview class with the name
skyview.geometry.projecter.Ait, though the comments in that file do
describe it as Hammer-Aitoff.  As Sergei Klioner, who pointed this out,
says:

  "If you use the Hammer-Aitoff projection, but call this Aitoff, it is a
   notation flaw which seem to have long tradition in astronomy (the reasons
   are unknown). For example, the projection used for Hipparcos books is the
   Hammer-Aitoff (see the given formulas - vol. 1, p 311), but is called Aitoff
   in the documentation."

So this long-standing topcat/stilts documentation error was at least
in good company.