ttools: avoid use of HumanMatchEngine
HumanMatchEngine was a wrapper class for MatchEngines that converted
tuple elements and match parameters into "human-friendly" units;
in practice this meant replacing radians with either degrees or arcseconds
depending on context.  It worked fine, but because of the MatchEngine
interface it had to operate by creating a lot of new objects at match
time.  Profiling determined that this was a major cost of resources
during large matches, increasing object allocation and requiring
significantly more effort from the, already overworked, Garbage Collector,
which especially inhibited performance improvements when run in parallel.

This commit avoids use of HumanMatchEngine altogether in favour of
using instead specific human-friendly subclasses of the various
MatchEngine implementations.  This effects a change of units with
no associated object creation cost and is therefore much more
efficient in terms of object allocation.

The HumanMatchEngine class is retired to the testcases directory,
where parts of the implementation are just used to check that the
MatchEngines actually used by STILTS do in fact have human-friendly
units.

The behaviour following this commit should be faster but otherwise
exactly the same, with one exception: custom MatchEngines supplied
by classname at runtime which have radian-unit tuples or match
parameters will no longer be modified to use human-friendly units.
But I bet nobody ever did that.

Actual profiling of the improved code definitely shows less pressure
on the Garbage Collector.  As far as overall elapsed time improvements
go, I think it is faster, but it doesn't seem as dramatic as examination
of changes to the CPU load on the GC would seem to suggest.

Note this only affects STILTS crossmatching, TOPCAT takes steps
to use the basic radians-based matchers already.