tjoin: refactor new interface MatchKit from MatchEngine
MatchEngine now acts as a factory for (factories of) MatchKit objects,
where MatchKit defines the compute-intensive parts of the
match-type-specific tuple-matching behaviour.

Unlike MatchEngines, MatchKit instances are intended to be used
only within a single thread, and so are able to work in a
non-threadsafe way, potentially permitting faster implementations
in some cases.

At this commit, this opportunity is taken to reuse some workspace
and the objects used for HEALPix calculations.  Profiling shows
the amount of memory allocation and CPU time associated with HEALPix
calculations is noticeably reduced as a result, but the total
runtime of multi-threaded match operations performed in tests
so far is not all that much improved.

Following this change, it would now be possible to make the HEALPix
calculations more efficient still by use of
cds.healpix.HealpixNested.newConeComputer4Xmatch (which FX assures
me is extremely fast) instead of the existing newConeComputerApprox,
but profiling seems to show that the HEALPix calculations are now
pretty insignificant in terms of memory allocation and CPU time
compared to other operations required by the matching, so there
doesn't seem much point.

The dominant users of CPU cycles for some example large matches
following this commit seem to be random access to scattered data inside
ByteBuffer implementations, and garbage collection.  Having played
around with these a bit, I don't see any obvious way to make
significant improvements.