fits: reduce FileChannel.map() calls
Add a new layer of ByteBuffer management (new classes BufferManager
and BlockManager) to be less extravagant with the number of actual
file mapping calls that are made.  Instead of calling FileChannel.map
every time a new ByteBuffer is required, try to limit the number of
maps to one per file region, and supply buffers using
ByteBuffer.duplicate().

This is perhaps a bit tidier than the previous code, though there is
an extra layer of code for keeping track of and releasing ByteBuffers.
That is slightly concerning since the unmapping is an unsafe operation,
but I don't *think* the changes have introduced new potential problems.

It's not really clear how bad the previous usage was (probably
OS-dependent), but reducing the number of mmap() system calls
is likely to be a good idea, and profiling seems to suggest that
this change reduces the number of page faults during multi-threaded
processing, so it may have a significant effect on performance.