fits: fix colfits string threading bug
The ColFitsStarTable.getRowAccess method was returning an object
which could give incorrect values for String- or String[]-valued
columns.  The problem was its use of a shared ValueReader object
which was not threadsafe (it's that premature optimisation again).
The effect was that in a multi-threaded context the string values
came out the right length, but with each character typically
sampled from different rows.  This effect could be quite nasty
in that it's not always easy to spot, e.g. for gaia_source
designation strings.  However RowAccess-based reads would not be
used all that much for strings, so impact in practice was probably
low. The main parallel string column operations on colfits files
affected that I can think of would be statistics (cardinality wrong,
min/max values garbled) and sorting (sort doesn't work;
fails with error, doesn't exit, gives wrong sequence).

Fix it so that there is no mutable workspace in ValueReader instances,
and document the ValueReader class (which pre-dates the multithreading
implementation) as thread-safe.