ttools: note NULL/NaN issues in user documentation
SUN/256 and SUN/253 now contain a note to the effect that users
should not attempt to distinguish between NULL and NaN values,
since STIL usually doesn't.

This information is in the section on JEL expressions; strictly
speaking its applicability is wider than that, but it's a reasonable
place that people might go looking for it.

Suggested by Berry Holl.