tjoin: add Coverage interface and implementations
Coverage is a geometry-agnostic abstraction that is able to
characterise coverage of a region of tuple-space.
It is intended as a generalisation and replacement of the NdRange
class, which effectively only understands Cartesian coordinates.

This commit introduces the Coverage interface, and implementations
suitable for Cartesian and Sky geometries.
No use of it is made as yet.