tjoin: RowMatcher uses Coverage not NdRange
NdRange is no longer used when matching, the more general Coverage
abstraction is used instead for assessing the regions of tuple space
in which table rows are present.

This is cleaner and allows more geometrically aware implementations,
in particular use of sky-specific coverage maps rather than simply
looking at ranges of longitude and latitude coordinates.

This ought to result in faster matches for some cases, especially
sky matches where the input tables cover partially disjoint regions
of the sky; now a better attempt can be made to avoid binning rows
that are known to be in regions where no potential matches exist.