util: modify missing class error reports
UTIL and STILTS both caught missing classes and advised that the
problem was probably an incomplete J2SE implementation in gcj.
This is no longer very likely, so modify or remove the relevant
warnings.  In the case of util.Loader it's not likely to happen,
and for stilts, it's most likely to be a misconfigured jar file or
something, in which case standard error reporting is reasonable
behaviour.