ttools: let ArrayQuantile plotter cope with missing array
If the X or Y array coordinate is missing for the ArrayQuantile plotter
it fills it in with a suitable sequence array (0, 1, 2, ... of the
right length) where appropriate.
At this commit it will never see such arrays, since the
FloatingArrayCoords that it uses are marked mandatory, but if that
is relaxed it means you can do plots without requiring index-only
arrays to be present.