ttools: accept missing X or Y array in ArrayShapePlotter
If either the Xs or the Ys coordinate array (but not both)
is missing for the 2d array-based ShapeForms
(MarkArrayForm, LineArrayForm, ErrorArrayForm, providing marks,
lines, yerrors and xyerrors plot layer types),
then the missing array will be interpreted as if it was a simple
sequence (0, 1, 2, ...) of the same length as the array that is present.

At present that won't happen, because FloatingArrayCoord.X/Y are
marked as mandatory, but if that restriction is lifted then
users will be able to plot array values without explicitly
supplying both X and Y arrays.