ttools: FloatingArrayCoord.X/Y are now optional
The array coordinates for X/Y array plots are no longer designated
mandatory for plots.  That means that either or both may be missing;
logic is now in place for the relevant plots to cope with that.
The point is so that users can e.g. omit the X array and get a
plot that works as if it were a sequence array (0,1,2,...) of the
size matching the Y array.

Examples and documentation are updated accordingly; in particular
prominent use of the sequence() function as an X array value is removed,
since users no longer need to be made aware of this somewhat clunky idiom.