fits: add sanitizeText option to CardFactory.Config
The FITS CardFactory now has a pluggable option to check free text
inputs (string values and comments, but not key names) and ensure
that they conform to the FITS constraints, i.e. all characters
are in the range 0x20-0x7e.  By default, it now replaces any
out-of-range characters with a question mark '?'.

The previous behaviour was to throw an IllegalArgumentException
if illegal FITS header characters were encountered in input
string values or comments.  That was pretty disastrous, since it
meant that non-ASCII Unicode characters from input table metadata
could (and did) cause the whole FITS output operation to fail.