ttools: re-implement array per-element arithmetic funcs
Rewrite the element-wise add, subtract, multiply and divide functions
in the func.Arrays class for use in the expression language.

Previously, these had two array parameters and in some cases
overloaded versions with one array and one scalar parameter,
in that order.
It turned out that the array+scalar versions didn't supply
additional type safety because primitive parameters are apparently
boxed by JEL to wrapper Objects if required, so for instance a
(double,Object) invocation would get dispatched to the (Object,Object)
overloaded method rather than producing a CompilationError if no
(double,Object) overloaded method existed.  So no additional harm
is done by removing the overloads and giving them all (Object,Object)
signatures instead - it would be a major pain to do it for specific
types because all the different numeric array types and combinations
would have to be declared separately.

So all these functions now exist in only one form with signature
double[] op(Object arrayOrScalar, Object arrayOrScalar), and can
take either argument (though not both) as a scalar instead of an array.
This gives more functionality for the user who now (a) can use scalars
for subtract and divide functions (previously that only worked for
add and multiply), and (b) doesn't have to remember if the argument
sequence is (scalar,array) or (array,scalar).