pds: don't try to parse base16/8/2 as numeric
The types ASCII_NUMERIC_BASE16, ASCII_NUMERIC_BASE8 and
ASCII_NUMERIC_BASE2 cannot actually be parsed as numeric values;
at least for the case of BASE16 I think this is a bug in the
pds4-jparser classes that I've copied into this package.
At this commit, just stop trying to parse them as such,
which will allow tables with such fields to be read without
failure.