vo: use optional nrows column from TAP_SCHEMA
If the optional, semi-standard "nrows" column is present in the
TAP_SCHEMA.tables table, use it to report the (estimated) row count
in TAP table metadata.  This is already got from the /tables endpoint
using the VODataService 1.2 "nrows" attribute, but till now was
not available for tap_schema metadata acquisition.
nrows in tap_schema is not so far in any standard, but it's marked
for probable inclusion in some future TAP update.