starjava: improve logging suppression during unit tests
Issuing log messages during unit tests generally just clutters up
the output, so it should be avoided.  Go through most/all of the
packages and fix it so that this is actually the case.

This is mainly done by replacing the use of
Logger.getLogger(name).setLevel(level) with the drop-in equivalent
from util.LogUtils, which is not at the mercy of the garbage collector.
A bit of other related tidying got done at the same time.