xdoc: allow links to non-section elements
Add id attributes to some elements (p, dt, li, figure) that do not
correspond to section-like files in the multi-file HTML output,
and update the XSLT processing so that links to these elements work.
That means allowing links to reference fragments within pages
in certain circumstances.

I had some trouble as usual with the XSLT.  I believe it's correct
and working, but there is repeated code (a long ancestor-or-self
disjunction xpath) to determine what (sub)section a node is in.
This should clearly be factored out into a separate ... template
or something, but I can't work out how to do it.  I'm sure it's
very easy.