topcat: adjust DataGeom matching re-range rules
Modify the criteria for when a plot re-ranging operation is forced.
It was triggered when the DataGeoms didn't match between layers,
but in some cases (the soon-to-appear HandleArrayForm)
the geom can change without a re-range being desirable,
though in some case (SkyDataGeom changing projection) it is.

So introduce a hack that allows both behaviours: if the getVariantName
has changed between DataGeoms, it counts as a big enough change to
warrant a re-range, otherwise not.