topcat: delegate maskId assignment to RowSubsets
The maskId used by GuiDataSpec to determine whether RowSubsets
have changed is now supplied by the RowSubset itself,
rather than determined by the GuiDataSpec based on the identity
of the RowSubset.

That means that a RowSubset instance is able to signal to the plotting
system, by changing its maskId, that a replot is required.

Note that using mutable state here is a bit more tricksy than it
looks at first sight, because of how and when the various plot
state identities are acquired and cached by PlotPanel to determine
replot requirements, which is where all this stuff gets used.
But I think these changes will generally work as the plotting
system is currently set up.