topcat: TopcatEvent.ROW now fires for activation cleared
TopcatListeners are now notified when the activation has been
cleared as well as when a new row is activated, with a
TopcatEvent.ROW message having a null datum.
This event can be triggered by calling TopcatModel.highlightRow(-1).

The only class to trigger such events is currently StackPlotWindow,
when the user clicks on a part of the plot near no suitable points.