topcat: identify RowSubsets by key
Add a getKey method to RowSubset to be used when identifying
RowSubsets by the GUI, for instance so that stored configuration
can be applied.  This allows subsets to change name or change
definition and still e.g. keep the same colour and position in
plot subset stacks.

This is a usability improvement, since for instance now when you
redefine a subset by drawing a blob, it no longer gets a new
colour and position in the stack.