ttools: don't build SUN/256 pdf by default
The docs (hence install) target in the TTOOLS build.xml no longer
builds the PDF version of SUN/256 alongside the HTML version.
A new target docs-pdf is introduced which does that if you want it.

Note that the docs-pdf target now does not work in a headless
environment.  That's not essential at this commit, but future
build operations may require non-headless operation.