ttools: add global items to Plot2Task configs
Fix AbstractPlot2Task so that the ConfigMap it passes to plot layers
contains global items such as surface profile and captioner keys
as well as those explicitly requested by the PlotLayer.
In most cases these are not used, but (a) this matches more closely
what topcat does and (b) a few layer types do use some of this
information.  In particular SkyGrid uses things like the captioner
and grid configuration (as noted in its user documentation).
I haven't done a thorough check that all the global information
is now passed from AbstractPlot2Task to the layers.

Modify the SkyGrid example plot at this commit as well to provide
a test of the improved behaviour.  If the code fix is not applied,
the sexagesimal setting will not honoured in the SkyGrid layer.