tjoin: speed up RowLink.compareTo
The RowLink.compareTo method involved array allocation and so was
quite slow; this can slow things down noticeably since a sort is
typically required in match post-processing.

Add code that identifies a common special case and treats it in
a much more efficient way.  This speeds up the sort significantly
(factor of 6?).