tjoin: rework GroupID reassignment
MatchStarTables.makeJoinTable identifies group IDs for multiply-linked
rows, and then does some extra work to turn these into more
human-friendly values (lower numbers) for, so far, only cosmetic reasons.
This commit changes the way that this is done, so that this ID remapping
is performed just before the output table is constructed rather than
in the (earlier) output of the findGroups method.

The result is the same (table output for match tables with Group IDs
should be identical to the previous behaviour) but this change has
some benefits.
First, it seems to be faster - both require a new Hash creation, but
now a sort and a load of binary searches are avoided.
Second, it's more robust against changes to the way that the
RowRef->Token map is put together, in particular it will produce
the same output regardless of the GroupID values actually used
in the Tokens.  That doesn't matter at the moment, since that map
is currently assembled in a deterministic way, but it may be useful
to preserve output unchanged (e.g. for regression tests) if earlier
parts of the algorithm become multithreaded and hence less deterministic.