tjoin: ProgressIndicator is now thread-safe
The ProgressIndicator.setLevel() method is being called from multiple
threads concurrently, e.g. from ParallelMatchComputer.progressCollect,
in contravention of the thread-safety rules in ProgressIndicator's
public contract, which says it should only be used from a single thread.

Fix this by changing the public contract of ProgressIndicator,
requiring implementations to be thread-safe.  This requires a change to
TextProgressIndicator (use an AtomicInteger); RangeModelProgressIndicator
is already thread-safe since it queues update behaviour on the EDT.

This fixes an issue that meant the text progress got confused and
sometimes printed too many dots on a line when running in parallel.
Now the dots should all line up.