tjoin: wrap sequential ProgressIndicator in ProgressTracker
New class ProgressTracker provides a wrapper for ProgressIndicator
where it is used in a sequential context.  This simplifies
ProgressIndicator use a little bit, but more importantly it
manages making ProgressIndicator.setLevel calls at a reasonable
frequency, i.e. not every iteration, since setLevel may not have
a negligable cost.  Use this ProgressTracker where applicable,
which is mostly in RowMatcher.