tjoin: MatchStarTables now configurable
The MatchStarTables class, which used to contain only static methods,
now has a constructor that takes a ProgressIndicator and SplitProcessor,
and some of its methods are now instance methods that may need to
use those members.

That means that the operations in MatchStarTables, which generally
take place after the match proper, and some of which may be
time-consuming, can now be logged and/or multithreaded,
which will make it easier to identify and fix bottlenecks.

At this commit, the MatchStarTables instance methods do not in
fact make use of those members.