tjoin: use Collectors in MatchStarTables
Rework some of the code in the MatchStarTables class to use a
SplitProcessor via the CollectionRunner class.
This makes it a one-liner to run across multiple threads parts
of the crossmatching post-processing that are at least in some
cases time-consuming.

However, tests show that this at least sometimes slows things down
rather than speeding them up, probably to do with having to do more
HashMap manipulation in the merge part of the parallel processing(?).
Where this seems to be the case, leave the code in place for possible
future use or experimentation, but for now keep the actual
implementation single-threaded.

Behaviour should not change.