util: use lazy ForkJoinPool acquisition in SplitPolicy
A custom ForkJoinPool is now configured in a SplitPolicy by
passing at construction time a Supplier<ForkJoinPool> rather than
a ForkJoinPool itself.  That means that FJP acquisition can be lazy,
so that a SplitPolicy can be created without creating a ForkJoinPool
(which may consume resources) if it never gets used.