topcat: docs-pdf build target not done headless
Building the PDF version of SUN/253 fails in headless mode,
since adding SVGs to it (for reasons I haven't tried very hard
to understand).  I realised this and took appropriate measures
for the TTOOLS build when I added it at 0af827d92, but omitted
to do the same thing for TOPCAT at 79e13e3ff.