tjoin: use long not int for Cartesian axis bins
AbstractCartesianMatchEngine and its concrete subclasses chop the
axes into bins and label each bin with an integer.
In the case of bins that are very small compared to the range of
values on the axis (real case: OBMT values in the range 0-10000
with matching tolerance ~1e-10), this could lead to integer overflow
so that many bins ended up with a value of Integer.MAX_VALUE or
Integer.MIN_VALUE (this value is calcluated using (int) Math.floor(...)).
This is not a correctness error, but either or both of those extremal
bins could end up very full, which would lead to matching so slow it
would probably never terminate.

Improve matters by using longs not ints to label the bins.
This is much less likely to run out of bins; it's still possible
and would again result in a performance issue or likely match failure
rather than a correctness error.

This approach will use more memory for bin keys, but not much,
since there is the overhead in any case of at least one Object
(actually two as currently implemented: the Cell object and the
long[] array it contains) when storing the primitive values.