ttools: no zoom for zero-extent range in SinProjection
When plotting a single point on the sky, don't zoom in to maximum
magnification (which is currently a few milliarcsec), since that's
arbitrary and works badly for e.g. the Area plot layer.
Instead zoom all the way out and show the whole sky, but still
centered on the relevant position.

This doesn't affect the other (all-sky) projections.
The Cartesian-based plots (Plane, Cube) make their own arrangements
for this case, using a zoom of approximately unit size.