votable: modify FIELD/LINK output
Modify behaviour when writing URL-valued aux metadata items in
FIELD elements: a ValueInfo for which getName() returns null
now results in a LINK element with no title attribute, rather than
as before suppressing the output of the LINK element altogether.

Normal usage of this code is not likely to notice the difference,
since ValueInfos would normally always have a name.

However, this change was requested by Gregory Mantelet at CDS
to support writing FIELD/LINK output in a way that is interpreted
by Aladin in a particular idiosyncratic way (in turn necessitated
by the fact that the LINK attributes are woefully underdocumented
in the VOTable standard).