table: cope with JDBC Array values
Values read by JDBC from the database as java.sql.Array types were
previously presented as such in input tables; these Array objects
cannot in general be written to output tables or otherwise manipulated
usefully by STIL/STILTS/TOPCAT.

This commit attempts to read the data as a STIL-friendly array type,
namely String[] or a primitive array type.
Some guesswork is required, since while the ResultSetMetaData does
report the type as an Array, it doesn't say what it's an array of.
So we dig into the ResultSetMetaData.getColumnTypeName to try to
work that out.  Minimal testing (float[] data from PostgreSQL)
confirms that it works at least some of the time.